De-angularize manifest
