+++ src/app/services/manifest-service.ts
@@ -1,0 +1,259 @@
+import * as _ from 'underscore';
+import * as idbKeyval from 'idb-keyval';
+
+// For zip
+// tslint:disable-next-line:no-implicit-dependencies
+import 'imports-loader?this=>window!@destiny-item-manager/zip.js';
+// tslint:disable-next-line:no-implicit-dependencies
+import inflate from 'file-loader?name=[name]-[hash:6].[ext]!@destiny-item-manager/zip.js/WebContent/inflate.js';
+// tslint:disable-next-line:no-implicit-dependencies
+import zipWorker from 'file-loader?name=[name]-[hash:6].[ext]!@destiny-item-manager/zip.js/WebContent/z-worker.js';
+
+import { requireSqlLib } from './database';
+import { reportException } from '../exceptions';
+import { getManifest as d2GetManifest } from '../bungie-api/destiny2-api';
+import { getManifest as d1GetManifest } from '../bungie-api/destiny1-api';
+import { IPromise } from 'angular';
+import { settings } from '../settings/settings';
+import { $q, $http, $rootScope } from 'ngimport';
+import { toaster } from '../ngimport-more';
+import { t } from 'i18next';
+import { DestinyManifest } from 'bungie-api-ts/destiny2';
+
+declare const zip: any;
+
+// This file exports D1ManifestService and D2ManifestService at the bottom of the
+// file (TS wants us to declare classes before using them)!
+
+// Testing flags
+const alwaysLoadRemote = false;
+
+class ManifestService {
+  isLoaded = false;
+  isError = false;
+  statusText: string | null = null;
+  version: string | null = null;
+
+  /**
+   * This tells users to reload the extension. It fires no more
+   * often than every 10 seconds, and only warns if the manifest
+   * version has actually changed.
+   */
+  warnMissingDefinition = _.debounce(() => {
+    this.getManifestApi()
+      .then((data) => {
+        const language = settings.language;
+        const path = data.mobileWorldContentPaths[language] || data.mobileWorldContentPaths.en;
+
+        // The manifest has updated!
+        if (path !== this.version) {
+          toaster.pop('warning',
+                      t('Manifest.Outdated'),
+                      t('Manifest.OutdatedExplanation'));
+        }
+      });
+  }, 10000, true);
+
+  private manifestPromise: Promise<any> | null = null;
+  private makeStatement = _.memoize((table, db) => {
+    return db.prepare(`select json from ${table} where id = ?`);
+  });
+
+  constructor(
+    readonly localStorageKey: string,
+    readonly idbKey: string,
+    readonly getManifestApi: () => IPromise<DestinyManifest>
+  ) {}
+
+  // TODO: redo all this with rxjs
+  getManifest() {
+    if (this.manifestPromise) {
+      return this.manifestPromise;
+    }
+
+    this.isLoaded = false;
+
+    this.manifestPromise = Promise
+      .all([
+        requireSqlLib(), // load in the sql.js library
+        this.loadManifest()
+      ])
+      .then(([SQLLib, typedArray]) => {
+        this.statusText = `${t('Manifest.Build')}...`;
+        const db = new SQLLib.Database(typedArray);
+        // do a small request, just to test it out
+        this.getAllRecords(db, 'DestinyRaceDefinition');
+        return db;
+      })
+      .catch((e) => {
+        let message = e.message || e;
+        this.statusText = t('Manifest.Error', { error: message });
+
+        if (e.status === -1) {
+          message = navigator.onLine
+            ? t('Bungiethis.NotConnectedOrBlocked')
+            : t('Bungiethis.NotConnected');
+        // tslint:disable-next-line:space-in-parens
+        } else if (e.status === 503 || e.status === 522 /* cloudflare */) {
+          message = t('Bungiethis.Down');
+        } else if (e.status < 200 || e.status >= 400) {
+          message = t('Bungiethis.NetworkError', {
+            status: e.status,
+            statusText: e.statusText
+          });
+        } else {
+          // Something may be wrong with the manifest
+          this.deleteManifestFile();
+        }
+
+        this.manifestPromise = null;
+        this.isError = true;
+        console.error("Manifest loading error", { error: e }, e);
+        reportException('manifest load', e);
+        throw new Error(message);
+      });
+
+    return this.manifestPromise;
+  }
+
+  getRecord(db, table, id): object | null {
+    const statement = this.makeStatement(table, db);
+    // The ID in sqlite is a signed 32-bit int, while the id we
+    // use is unsigned, so we must convert
+    const sqlId = new Int32Array([id])[0];
+    const result = statement.get([sqlId]);
+    statement.reset();
+    if (result.length) {
+      return JSON.parse(result[0]);
+    }
+    return null;
+  }
+
+  getAllRecords(db, table): object {
+    const rows = db.exec(`SELECT json FROM ${table}`);
+    const result = {};
+    rows[0].values.forEach((row) => {
+      const obj = JSON.parse(row);
+      result[obj.hash] = obj;
+    });
+    return result;
+  }
+
+  private loadManifest(): IPromise<Uint8Array> {
+    return $q.all([
+      this.getManifestApi(),
+      settings.ready // wait for settings to be ready
+    ])
+      .then(([data]) => {
+        const language = settings.language;
+        const path = data.mobileWorldContentPaths[language] || data.mobileWorldContentPaths.en;
+
+        // Use the path as the version, rather than the "version" field, because
+        // Bungie can update the manifest file without changing that version.
+        const version = path;
+        this.version = version;
+
+        return this.loadManifestFromCache(version)
+          .catch(() => this.loadManifestRemote(version, path));
+      });
+  }
+
+  /**
+   * Returns a promise for the manifest data as a Uint8Array. Will cache it on succcess.
+   */
+  private loadManifestRemote(version, path): IPromise<Uint8Array> {
+    this.statusText = `${t('Manifest.Download')}...`;
+
+    return $http.get(`https://www.bungie.net${path}?host=${window.location.hostname}`, { responseType: "blob" })
+      .then((response) => {
+        this.statusText = `${t('Manifest.Unzip')}...`;
+        return unzipManifest(response.data);
+      })
+      .then((arraybuffer) => {
+        this.statusText = `${t('Manifest.Save')}...`;
+
+        const typedArray = new Uint8Array(arraybuffer);
+        // We intentionally don't wait on this promise
+        idbKeyval.set(this.idbKey, typedArray)
+          .then(() => {
+            console.log(`Sucessfully stored ${typedArray.length} byte manifest file.`);
+            localStorage.setItem(this.localStorageKey, version);
+          })
+          .then(null, (e) => {
+            console.error('Error saving manifest file', e);
+            toaster.pop({
+              title: t('Help.NoStorage'),
+              body: t('Help.NoStorageMessage'),
+              type: 'error'
+            }, 0);
+          });
+
+        $rootScope.$broadcast('dim-new-manifest');
+        return typedArray;
+      });
+  }
+
+  private deleteManifestFile() {
+    localStorage.removeItem(this.localStorageKey);
+    idbKeyval.delete(this.idbKey);
+  }
+
+  /**
+   * Returns a promise for the cached manifest of the specified
+   * version as a Uint8Array, or rejects.
+   */
+  private loadManifestFromCache(version): IPromise<Uint8Array> {
+    if (alwaysLoadRemote) {
+      throw new Error("Testing - always load remote");
+    }
+
+    this.statusText = `${t('Manifest.Load')}...`;
+    const currentManifestVersion = localStorage.getItem(this.localStorageKey);
+    if (currentManifestVersion === version) {
+      return $q.when(idbKeyval.get(this.idbKey)).then((typedArray: Uint8Array) => {
+        if (!typedArray) {
+          throw new Error("Empty cached manifest file");
+        }
+        return typedArray;
+      });
+    } else {
+      ga('send', 'event', 'Manifest', 'Need New Manifest');
+      throw new Error(`version mismatch: ${version} ${currentManifestVersion}`);
+    }
+  }
+}
+
+/**
+ * Unzip a file from a ZIP Blob into an ArrayBuffer. Returns a promise.
+ */
+function unzipManifest(blob): IPromise<ArrayBuffer> {
+  return $q((resolve, reject) => {
+    zip.useWebWorkers = true;
+    zip.workerScripts = {
+      inflater: [zipWorker, inflate]
+    };
+    zip.createReader(new zip.BlobReader(blob), (zipReader) => {
+      // get all entries from the zip
+      zipReader.getEntries((entries) => {
+        if (entries.length) {
+          entries[0].getData(new zip.BlobWriter(), (blob) => {
+            const blobReader = new FileReader();
+            blobReader.addEventListener("error", (e) => { reject(e); });
+            blobReader.addEventListener("load", () => {
+              zipReader.close(() => {
+                resolve(blobReader.result);
+              });
+            });
+            blobReader.readAsArrayBuffer(blob);
+          });
+        }
+      });
+    }, (error) => {
+      reject(error);
+    });
+  });
+}
+
+// Two separate copies of the service, with separate state and separate storage
+export const D1ManifestService = new ManifestService('manifest-version', 'dimManifest', d1GetManifest);
+export const D2ManifestService = new ManifestService('d2-manifest-version', 'd2-manifest', d2GetManifest);
