+++ src/app/settings/settings.module.ts
@@ -1,12 +1,11 @@
 import { module } from 'angular';
 
-import { SettingsComponent, SettingsController } from './settings.component';
+import { SettingsComponent } from './settings.component';
 import SortOrderEditor from './sort-order-editor';
 import { react2angular } from 'react2angular';
 import { settings } from '../settings/settings';
 
 export default module('settingsModule', [])
-  .controller('dimSettingsCtrl', SettingsController)
   .component('settings', SettingsComponent)
   .component('sortOrderEditor', react2angular(SortOrderEditor, ['order', 'onSortOrderChanged']))
   .config(($stateProvider) => {
