+++ src/app/shell/dimManifestProgress.directive.js
@@ -1,11 +1,12 @@
-import angular from 'angular';
+import { module } from 'angular';
+import { D1ManifestService, D2ManifestService } from '../services/manifest-service';
 import template from './dimManifestProgress.directive.html';
 import './dimManifestProgress.scss';
 
 /**
  * A dialog that shows the progress of loading the manifest.
  */
-angular.module('dimApp')
+module('dimApp')
   .component('dimManifestProgress', {
     template,
     controller: ManifestProgressCtrl,
@@ -14,10 +15,6 @@ angular.module('dimApp')
     }
   });
 
-function ManifestProgressCtrl(dimManifestService, D2ManifestService) {
-  if (this.destinyVersion === 2) {
-    this.manifest = D2ManifestService;
-  } else {
-    this.manifest = dimManifestService;
-  }
+function ManifestProgressCtrl() {
+  this.manifest = this.destinyVersion === 2 ? D2ManifestService : D1ManifestService;
 }
