+++ src/index.js
@@ -7,7 +7,6 @@ import './app/exceptions';
 // Initialize the main DIM app
 import './app/app.module';
 
-import './app/services/dimManifestService.factory';
 import './app/services/dimInfoService.factory';
 import './app/services/dimDestinyTrackerService.factory';
